miner/pyproject.toml
[tool.poetry]
name = "miner"
version = "0.1.0"
description = ""
authors = ["Christian Birchler <christian.birchler2@uzh.ch>"]

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.55"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
